// kodikosERP-master/src/app/models/formatki/offerform.model.ts
export interface Offerform {
  data?: boolean;
  creator?: boolean;
  termsOfPayment?: boolean;
  deliveryType?: boolean;
  labelling?: boolean;
  packing?: boolean;
  certificate?: boolean;
  realization?: boolean;
  engeneeringTolerance?: boolean;
  origin?: boolean;
  comments?: boolean;
  contractor?: boolean;
  addDate?: boolean;
  modDate?: boolean;
  addBy?: boolean;
  modBy?: boolean;
  items?: boolean;
  contractorShortName?: boolean;
  contractorFullName?: boolean;
  contractorNIP?: boolean;
  currency?: boolean;
  project?: boolean;
  company?: boolean;
}